added constants.py, utils.py, and refactored link_checker.py in order to support checking deeds and rdf
